refactor to use Builder
