+++ src/Illuminate/Database/Schema/PostgresBuilder.php
@@ -5,6 +5,26 @@ namespace Illuminate\Database\Schema;
 class PostgresBuilder extends Builder
 {
     /**
+     * Drop all tables from the database.
+     *
+     * @return void
+     */
+    public function dropAllTables()
+    {
+         $tables = [];
+
+         foreach ($this->connection->select($this->grammar->compileGetAllTables($this->connection->getConfig('schema'))) as $table) {
+            $tables[] = get_object_vars($table)[key($table)];
+         }
+
+         if (empty($tables)) {
+            return;
+         }
+
+         $this->connection->statement($this->grammar->compileDropAllTables($tables));
+    }
+
+    /**
      * Determine if the given table exists.
      *
      * @param  string  $table
